Create `fake_learner_profiles` table (#4)
* Enable MODDATETIME extension for timestamps

See https://dev.to/paullaros/updating-timestamps-automatically-in-supabase-5f5o

* Add migration + seed for `fake_learner_profiles`

Also adding a `db:reset` script which drops the DB, recreates it, runs
migrations, and seeds.

* Add basic DB test for `fake_learner_profiles`

The process of writing this was... not fun. I was attempting more of a
behavioral test for the timestamp trigger, but I couldn't get it to
work.