Install Sentry (#10)
* Add coverage directories to eslint's ignore-list

* Install Sentry

This setup was generated by the Sentry wizard; i.e.

`npx @sentry/wizard@latest -i nextjs`

* Add test page for Sentry errors

This was also generated by the Sentry wizard, but I've opted to commit
it separately so that we can eventually remove it. For now though, it
will be useful as we tweak our Sentry configuration to get the details
we need in error reports.

* Require newline at the end of files