Connect course outline edits to Supabase (#45)
* Connect initial edit load & form state to backend

Save doesn't yet work though, still need to update that endpoint.
Removed some files that are no longer used.

* Rename `useCourseOutlineData` to `useCourseOutlineEdit`

Also made the ID a param. I might be wrong here, but it felt more like
the Page component's responsibility to read params.

* Use Supabase for course outline updates

Also introduced Zod for JSON schema validation on the endpoint.

* Rename `fetchers` to `services`

* Rename `useCourseOutlinesList` to `useCourseOutlines`

This is a bit more consistent with the individual hook,
`useCourseOutline`.

* Extract service for updating a course outline

This also introduces a pattern where the model knows how to serialize
itself as update params.

* Rename edit hook